refactor: move browser-manager to manager folder to organaize code and seperate concerns
