+++ packages/shortest/src/browser/browser.ts
@@ -15,7 +15,7 @@ import { writeFileSync, mkdirSync } from 'fs';
 import { rm } from 'fs/promises';
 import { join } from 'path';
 import { GitHubTool } from './integrations/github';
-import { BrowserManager } from './browser-manager';
+import { BrowserManager } from './manager';
 import { TestContext, BrowserToolConfig } from '../types/index';
 
 export class BrowserTool extends BaseBrowserTool {
